[MOV] *: move initAutoMoreMenu to a non-lazy loaded file
*: web, website

task-2618929
